# openapi.yaml
openapi: 3.0.3
info:
  description: "API for managing agents, teams, and tasks in an AutoGen system."
  title: AutoGen Framework API
  version: 1.0.0
servers:
- description: Production environment
  url: https://api.autogenframework.com/v1
- description: Local development environment
  url: http://localhost:3000
paths:
...
  /teams:
    get:
      description: Retrieves a list of all teams in the system.
      operationId: listTeams
      parameters:
      - description: Maximum number of teams to return.
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          minimum: 1
          type: integer
        style: form
      - description: Pagination offset.
        explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Team'
                type: array
          description: A list of teams
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listAgents_default_response'
          description: An error response
      summary: List all teams
      tags:
      - Teams
    post:
      description: Adds a new team to the system.
      operationId: createTeam
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamInput'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
          description: Team created successfully
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listAgents_default_response'
          description: An error response
      summary: Create a new team
      tags:
      - Teams
  /teams/{team_id}:
    delete:
      description: Removes a team from the system by its ID.
      operationId: deleteTeam
      parameters:
      - description: The ID of the team
        explode: false
        in: path
        name: team_id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          description: Team deleted successfully
        "404":
          description: Team not found
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listAgents_default_response'
          description: An error response
      summary: Delete an existing team
      tags:
      - Teams
    get:
      description: Returns details about a particular team by its ID.
      operationId: getTeamById
      parameters:
      - description: The ID of the team
        explode: false
        in: path
        name: team_id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
          description: Team details
        "404":
          description: Team not found
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listAgents_default_response'
          description: An error response
      summary: Retrieve a specific team
      tags:
      - Teams
    patch:
      description: Allows partial updates to an existing team’s information.
      operationId: updateTeam
      parameters:
      - description: The ID of the team
        explode: false
        in: path
        name: team_id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamUpdate'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
          description: Team updated successfully
        "404":
          description: Team not found
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listAgents_default_response'
          description: An error response
      summary: Update an existing team
      tags:
      - Teams
...
components:
...
  schemas:
...
    Team:
      example:
        name: name
        description: description
        created_at: 2000-01-23T04:56:07.000+00:00
        team_id: 0
      properties:
        team_id:
          description: Unique identifier of the team
          title: team_id
          type: integer
        name:
          description: The name of the team
          title: name
          type: string
        description:
          description: A short description of the team
          title: description
          type: string
        created_at:
          description: Timestamp when the team was created
          format: date-time
          title: created_at
          type: string
      title: Team
      type: object
    TeamInput:
      example:
        name: name
        description: description
      properties:
        name:
          description: The name of the team
          title: name
          type: string
        description:
          description: A short description of the team
          title: description
          type: string
      required:
      - name
      title: TeamInput
      type: object
    TeamUpdate:
      description: Schema for partially updating a team
      example:
        name: name
        description: description
      properties:
        name:
          title: name
          type: string
        description:
          title: description
          type: string
      title: TeamUpdate
      type: object
...
    listAgents_default_response:
      example:
        code: code
        details: "{}"
        message: message
      properties:
        code:
          description: "A unique error code (e.g., \"AGENT_NOT_FOUND\")"
          title: code
          type: string
        message:
          description: A human-readable error message
          title: message
          type: string
        details:
          description: Additional context or fields related to the error
          title: details
          type: object
      required:
      - code
      - message
      title: listAgents_default_response
      type: object